kafka_consumer/requirements.txt
confluent-kafka==2.3.0
numpy==1.21.2
flask
werkzeug
requests
psycopg2-binary
